requirements.txt
aiogram==3.15.0
aiohttp==3.10.11
aiosqlite==0.20.0
APScheduler==3.10.4
cryptography==44.0.0
jdatetime==5.0.0
prometheus-client==0.21.1
python-dotenv==1.0.1
qrcode==7.4.2
Pillow==10.4.0
